handle delete cell by type

+++ src/input.js
@@ -24,6 +24,7 @@ import {
   deleteTable,
   selectedRect,
 } from './commands';
+import {columnTypesMap} from './columnsTypes/types.config';
 
 export const handleKeyDown = keydownHandler({
   ArrowLeft: arrow('horiz', -1),
@@ -55,7 +56,7 @@ function checkIfParentIsCell(state) {
   // if parent is not a table cell - let the editor handle key down
   if (parent.type.name !== 'table_cell') return false;
 
-  return true
+  return true;
 }
 
 function splitIfCellChild(state, dispatch) {
@@ -275,6 +276,25 @@ function shiftArrow(axis, dir) {
   };
 }
 
+export const deleteCellContentByType = (state, dispatch, view) => {
+  const {selection: sel, tr} = state;
+  if (!(sel instanceof CellSelection)) return false;
+
+  const reversedCellsInSelection = [];
+  sel.forEachCell((cell, pos) => {
+    reversedCellsInSelection.unshift({cell, pos});
+  });
+
+  reversedCellsInSelection.forEach(({cell, pos}) => {
+    const cellDeleteCommand =
+      columnTypesMap[cell.attrs.type].deleteContentCommand;
+    cellDeleteCommand(cell, pos, tr);
+  });
+
+  dispatch(tr);
+  return true;
+};
+
 export function getDeleteCommand(state) {
   if (!(state.selection instanceof CellSelection)) return null;
 
@@ -291,7 +311,7 @@ export function getDeleteCommand(state) {
   if (state.selection.isRowSelection()) return deleteRow;
   if (state.selection.isColSelection()) return deleteColumn;
 
-  return null;
+  return deleteCellContentByType;
 }
 
 function deleteCellSelection(state, dispatch) {
@@ -299,6 +319,7 @@ function deleteCellSelection(state, dispatch) {
   if (!(sel instanceof CellSelection)) return false;
   if (dispatch) {
     const deleteCommand = getDeleteCommand(state);
+    if (!deleteCommand) return false;
     return deleteCommand(state, dispatch);
   }
   return true;
@@ -429,12 +450,15 @@ export function handleMouseDown(view, startEvent) {
     if (anchor != null) {
       // Continuing an existing cross-cell selection
       $anchor = view.state.doc.resolve(anchor);
-    } else if (domInCell(view, event.target) && domInCell(view, event.target) != startDOMCell) {
+    } else if (
+      domInCell(view, event.target) &&
+      domInCell(view, event.target) != startDOMCell
+    ) {
       // Moving out of the initial cell -- start a new cell selection
       $anchor = cellUnderMouse(view, startEvent);
-      if (!$anchor){
+      if (!$anchor) {
         return stop();
-      } 
+      }
     }
     if ($anchor) setCellSelection($anchor, event);
